requirements.txt
certifi==2019.6.16
cffi==1.15.1
chardet==3.0.4
crypto==1.4.1
cryptography==39.0.2
fire==0.5.0
idna==2.8
image-slicer==0.2.0
lxml==4.9.2
Naked==0.1.31
numpy==1.16.4
opencv-python==4.1.0
panda==0.3.1
pdf2docx==0.5.1
Pillow==6.1.0
psutil==5.6.3
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.14.1
pygame==1.9.6
PyMuPDF==1.21.1
pyperclip==1.7.0
pypiwin32==223
python-docx==0.8.11
pywin32==305
PyYAML==5.1.1
requests==2.22.0
scapy==2.4.5
shellescape==3.4.1
six==1.12.0
termcolor==2.2.0
urllib3==1.25.3
wxPython==4.0.6
